Polished the MainManu element in the UI
When clicking the MainMenu button, a text appears to confirm if the player really wants to quit and warn that the run will end.

Before this change, after clicking the MainMenu button on the starting screen and then starting the level, when pause was triggered, the MainMenu element was already in the "exit confirmation state", instead of having the MainButton again. Same happened if this action was performed in the pause menu, after returning to gameplay and pausing again, the confirmation message was appearing instead of the button.

Now, that has been corrected and when the menu opens it will always show the MainMenu button.